+++ test/kubernetes/e2e/features/example/suite.go
@@ -38,14 +38,13 @@ func (s *FeatureSuite) BeforeTest(suiteName, testName string) {
 func (s *FeatureSuite) AfterTest(suiteName, testName string) {
 }
 
-// func (s *FeatureSuite) TestInstallationSuccessful() {
-// 	s.testInstallation.Assertions.EventuallyInstallationSucceeded(s.ctx)
-// }
-
-func (s *FeatureSuite) TestFailureAllowed() {
+func (s *FeatureSuite) TestExampleAssertion() {
 	// Testify assertion
 	s.Assert().NotEqual(1, 2, "1 does not equal 2")
 
-	// Gomega assertion
+	// Testify assertion, using the TestInstallation to provide it
+	s.testInstallation.Assertions.NotEqual(1, 2, "1 does not equal 2")
+
+	// Gomega assertion, using the TestInstallation to provide it
 	s.testInstallation.Assertions.Expect(1).NotTo(Equal(2), "1 does not equal 2")
 }
