--- test/kubernetes/e2e/test.go
@@ -1,16 +1,0 @@
-package e2e
-
-import (
-	"github.com/solo-io/gloo/test/kubernetes/testutils/cluster"
-	"github.com/solo-io/gloo/test/kubernetes/testutils/runtime"
-)
-
-func NewTestCluster() *TestCluster {
-	runtimeContext := runtime.NewContext()
-	clusterContext := cluster.MustKindContext(runtimeContext.ClusterName)
-
-	return &TestCluster{
-		RuntimeContext: runtimeContext,
-		ClusterContext: clusterContext,
-	}
-}
