+++ test/kubernetes/testutils/assertions/provider.go
@@ -1,9 +1,8 @@
 package assertions
 
 import (
+	"github.com/stretchr/testify/assert"
 	"testing"
-
-	"github.com/stretchr/testify/require"
 
 	"github.com/onsi/gomega"
 	"github.com/solo-io/gloo/test/kubernetes/testutils/gloogateway"
@@ -16,8 +15,12 @@ import (
 // So this provider maintains state about the install/cluster it is using, and then provides
 // operations.ClusterAssertion to match
 type Provider struct {
+	// We extend any assertions that are provided by testify/assert
+	*assert.Assertions
+
+	// Gomega is well-used around the codebase, so we also add support here
+	// NOTE TO DEVELOPERS: We recommend relying on testify assertions where possible
 	gomega.Gomega
-	*require.Assertions
 
 	clusterContext     *cluster.Context
 	glooGatewayContext *gloogateway.Context
@@ -27,8 +30,8 @@ type Provider struct {
 // installation of Gloo Gateway
 func NewProvider(t *testing.T) *Provider {
 	return &Provider{
+		Assertions: assert.New(t),
 		Gomega:     gomega.NewWithT(t),
-		Assertions: require.New(t),
 
 		clusterContext:     nil,
 		glooGatewayContext: nil,
@@ -51,5 +54,5 @@ func (p *Provider) WithGlooGatewayContext(ggCtx *gloogateway.Context) *Provider 
 // if the provider has been configured to point to a Gloo Gateway installation
 // There are certain Assertions that can be invoked that do not require that Gloo Gateway be installed for them to be invoked
 func (p *Provider) expectGlooGatewayContextDefined() {
-	p.Expect(p.glooGatewayContext).NotTo(gomega.BeNil(), "Provider attempted to create an Assertion that requires a Gloo Gateway installation, but none was configured")
+	p.Assertions.NotNil(p.glooGatewayContext, "Provider attempted to create an Assertion that requires a Gloo Gateway installation, but none was configured")
 }
